Deploy test/languageDetect/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==4.0.1
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
psycopg2==2.9.3
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.9
whitenoise==6.0.0
wincertstore==0.2
Pillow==8.2.0
detectlanguage